refactor: data layer
