+++ app/src/main/java/com/eventorium/data/auth/models/City.java
@@ -1,4 +1,4 @@
-package com.eventorium.data.models;
+package com.eventorium.data.auth.models;
 
 // NOTE: This class might have additional attributes in the future.
 // It will serve as a reference table in the database to prevent users from inventing city names.
