+++ app/src/main/java/com/eventorium/data/auth/models/Person.java
@@ -1,4 +1,4 @@
-package com.eventorium.data.models;
+package com.eventorium.data.auth.models;
 
 public class Person {
     private String name;
