+++ app/src/main/java/com/eventorium/data/auth/models/User.java
@@ -1,4 +1,4 @@
-package com.eventorium.data.models;
+package com.eventorium.data.auth.models;
 
 public class User {
     private String email;
