+++ app/src/main/java/com/eventorium/data/solution/models/Product.java
@@ -1,30 +1,29 @@
-package com.eventorium.data.models;
-
+package com.eventorium.data.solution.models;
 
 import android.os.Parcel;
 import android.os.Parcelable;
 
 import androidx.annotation.NonNull;
 
-public class Service implements Parcelable {
+public class Product implements Parcelable {
 
     private String name;
     private Double price;
-    private Integer photo;
+    public Integer photo;
 
-    public Service(String name, Double price, Integer photo) {
+    public String getName() { return name; }
+    public Double getPrice() { return price; }
+    public Integer getPhoto() { return photo; }
+
+    public Product(String name, Double price, Integer photo) {
         this.name = name;
         this.price = price;
         this.photo = photo;
     }
 
-    public Service() { }
+    public Product() {  }
 
-    public String getName() { return name; }
-    public Double getPrice() { return price; }
-    public Integer getPhoto(){ return photo; }
-
-    protected Service(Parcel in) {
+    protected Product(Parcel in) {
         name = in.readString();
         price = in.readDouble();
         photo = in.readInt();
@@ -40,12 +39,12 @@ public class Service implements Parcelable {
     @Override
     public int describeContents() { return 0; }
 
-    public static final Creator<Service> CREATOR = new Creator<>() {
+    public static final Creator<Product> CREATOR = new Creator<Product>() {
         @Override
-        public Service createFromParcel(Parcel in) { return new Service(in); }
+        public Product createFromParcel(Parcel in) { return new Product(in); }
 
         @Override
-        public Service[] newArray(int size) { return new Service[size]; }
+        public Product[] newArray(int size) { return new Product[size]; }
     };
-
 }
+
