+++ app/src/main/java/com/eventorium/presentation/category/adapters/CategoriesAdapter.java
@@ -10,7 +10,7 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.eventorium.R;
-import com.eventorium.data.models.Category;
+import com.eventorium.data.category.models.Category;
 import com.eventorium.presentation.util.listeners.OnEditClickListener;
 
 import java.util.List;
