+++ app/src/main/java/com/eventorium/presentation/category/adapters/CategoryProposalsAdapter.java
@@ -8,7 +8,7 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
-import com.eventorium.data.models.Category;
+import com.eventorium.data.category.models.Category;
 
 import com.eventorium.R;
 import com.eventorium.presentation.util.listeners.OnReviewProposalListener;
