+++ app/src/main/java/com/eventorium/presentation/solution/fragments/product/ProductOverviewFragment.java
@@ -11,7 +11,7 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.eventorium.R;
-import com.eventorium.data.models.Product;
+import com.eventorium.data.solution.models.Product;
 import com.eventorium.databinding.FragmentProductOverviewBinding;
 import com.eventorium.presentation.solution.adapters.ProductsAdapter;
